config: move config loading to a module
Then actually allow overriding it with an ENV var.
Previously, only one of the places it was loaded from respected the ENV
var.
