+++ .gitignore
@@ -1,5 +1,5 @@
 *.sw*
-config.js
+/config.js
 node_modules
 /views/standard/css
 pulldasher.pid
