+++ controllers/main.js
@@ -1,6 +1,6 @@
 var socketAuthenticator = require('../lib/socket-auth');
 var socketio = require('../package').dependencies['socket.io'];
-var config = require('../config');
+var config = require('../lib/config');
 
 module.exports = {
    index: function(req, res) {
