+++ lib/authentication.js
@@ -1,4 +1,4 @@
-var config = require('../config'),
+var config = require('./config'),
     debug = require('./debug')('pulldasher:authentication'),
     github = require('./git-manager').github,
     passport = require('passport'),
