+++ lib/config.js
@@ -1,0 +1,3 @@
+var configPath = process.env.CONFIG_PATH || '../config';
+module.exports = require(configPath);
+
