+++ lib/socket-auth.js
@@ -1,4 +1,4 @@
-var config = require('../config');
+var config = require('./config');
 var users = {};
 
 var token_timeout = config.token_timeout !== undefined ?
