+++ lib/utils.js
@@ -1,4 +1,4 @@
-var config = require('../config'),
+var config = require('./config'),
     Bluebird = require('bluebird'),
     Promise = global.Promise,
     _      = require('underscore');
